requirements.txt
beautifulsoup4===4.3.2
requests===2.4.1
getpass===1.6.2
Pillow===2.8.1
